Split up train and search code
It all lives in train.train and train.search now.
