+++ train.py
@@ -28,174 +28,17 @@ training run tensorboard from the log directory.
 """
 
 import argparse
-import datetime
-import logging
 import os
-import pickle
 
+import tensorflow as tf
 import matplotlib
 
 matplotlib.use("Agg")  # enable canvas drawing
 
-import tensorflow as tf
-
-from model_crnn import ModelCRNN_HQ, ModelCRNN_LQ
-from ml_tools.config import Config
-from ml_tools.dataset import dataset_db_path
 from ml_tools.logs import init_logging
-
-# this is a good list for a full search, but will take a long time to run (days)
-FULL_SEARCH_PARAMS = {
-    "batch_size": [1, 2, 4, 8, 16, 32, 64],
-    "learning_rate": [1e-3, 3e-4, 1e-4, 3e-5, 1e-5],
-    "l2_reg": [0, 1e-1, 1e-2, 1e-3, 1e-4, 1e-5, 1e-6, 1e-7, 1e-8],
-    "label_smoothing": [0, 0.01, 0.05, 0.1, 0.2],
-    "keep_prob": [0, 0.1, 0.2, 0.5, 0.8, 1.0],
-    "batch_norm": [True, False],
-    "lstm_units": [64, 128, 256, 512],
-    "enable_flow": [True, False],
-    "augmentation": [True, False],
-    "thermal_threshold": [-100, -20, -10, 0, 10, 20, 100],
-    # these runs will be identical in their parameters, it gives a feel for the varience during training.
-    "identical": [1, 2, 3, 4, 5],
-}
-
-# this checks just the important parameters, and only around areas that are likely to work well.
-# I've also excluded the default values as these do not need to be tested again.
-SHORT_SEARCH_PARAMS = {
-    "batch_size": [8, 32],
-    "l2_reg": [1e-2, 1e-3, 1e-4],
-    "label_smoothing": [0, 0.05, 0.2],
-    "keep_prob": [0.1, 0.4, 0.6, 1.0],
-    "batch_norm": [False],
-    "lstm_units": [128, 512],
-    "enable_flow": [False],
-    "augmentation": [False],
-    "thermal_threshold": [-100, -20, -10, 0, 20, 100],
-}
-
-
-def train_model(run_name, conf, hyper_params):
-    """Trains a model with the given hyper parameters.
-    """
-
-    run_name = os.path.join("train", run_name)
-
-    # a little bit of a pain, the model needs to know how many classes to classify during initialisation,
-    # but we don't load the dataset till after that, so we load it here just to count the number of labels...
-    dataset_filename = dataset_db_path(conf)
-    with open(dataset_filename, "rb") as f:
-        dsets = pickle.load(f)
-    labels = dsets[0].labels
-
-    model = ModelCRNN_LQ(labels=len(labels), **hyper_params)
-
-    model.import_dataset(dataset_filename)
-    model.log_dir = os.path.join(conf.train.train_dir, "logs")
-
-    # display the data set summary
-    print("Training on labels", labels)
-    print()
-    print(
-        "{:<20} {:<20} {:<20} {:<20} (segments/tracks/bins/weight)".format(
-            "label", "train", "validation", "test"
-        )
-    )
-    for label in labels:
-        print(
-            "{:<20} {:<20} {:<20} {:<20}".format(
-                label,
-                "{}/{}/{}/{:.1f}".format(*model.datasets.train.get_counts(label)),
-                "{}/{}/{}/{:.1f}".format(*model.datasets.validation.get_counts(label)),
-                "{}/{}/{}/{:.1f}".format(*model.datasets.test.get_counts(label)),
-            )
-        )
-    print()
-
-    for dataset in dsets:
-        print(dataset.labels)
-
-    print("Training started")
-    print("---------------------")
-    print("Hyper parameters")
-    print("---------------------")
-    print(model.hyperparams_string)
-    print()
-    print("Found {0:.1f}K training examples".format(model.rows / 1000))
-    print()
-    model.train_model(
-        epochs=conf.train.epochs,
-        run_name=run_name + " " + datetime.datetime.now().strftime("%Y%m%d-%H%M%S"),
-    )
-    model.save()
-    model.close()
-
-    # this shouldn't be nessesary, but unfortunately my model.close isn't cleaning up everything.
-    # I think it's because i'm adding everything to the default graph?
-    tf.reset_default_graph()
-
-    return model
-
-
-def axis_search(conf):
-    """
-    Evaluate each hyper-parameter individually against a reference.
-
-    The idea here is to assess each parameter individually while holding all other parameters at their default.
-    For optimal results this will need to be done multiple times, each time updating the defaults to their optimal
-    values.
-    """
-    logging.info("Performing hyper parameter search.")
-
-    results_filename = os.path.join(conf.train.train_dir, "search-results.txt")
-    logging.info(f"Writing hyper-parameter search results to {results_filename}")
-    tracker = JobTracker(results_filename)
-
-    # run the reference job with default params
-    run_job(tracker, "reference", conf)
-
-    for param_name, param_values in SHORT_SEARCH_PARAMS.items():
-        for param_value in param_values:
-            job_name = f"{param_name}={param_value}"
-            run_job(tracker, job_name, conf, {param_name: param_value})
-
-
-def run_job(tracker, job_name, conf, hyper_params=None):
-    """Run a job with given hyper parameters, and log its results.
-    """
-    if tracker.is_done(job_name):
-        return
-    if not hyper_params:
-        hyper_params = {}
-
-    print("-" * 60)
-    print("Processing", job_name)
-    print("-" * 60)
-
-    model = train_model(job_name, conf, hyper_params)
-    tracker.mark_done(job_name, model.eval_score, hyper_params)
-
-
-class JobTracker:
-    def __init__(self, filename):
-        self.filename = filename
-        open(self.filename, "w").close()  # Create/truncate
-
-    def is_done(self, job_name):
-        """Returns True if this job has been processed before.
-        """
-        with open(self.filename, "r") as f:
-            return any(line.split(",", 1)[0] == job_name for line in f)
-
-    def mark_done(self, job_name, score, hyper_params):
-        with open(self.filename, "a") as f:
-            f.write(
-                "{}, {}: {}\n".format(
-                    job_name,
-                    score,
-                    " ".join(f"{k}={v}" for (k, v) in hyper_params.items()),
-                )
-            )
+from ml_tools.config import Config
+from train.train import train_model
+from train.search import axis_search
 
 
 def load_config():
